+++ packages/react/src/router/v6v7/initialize.ts
@@ -1,6 +1,11 @@
 import hoistNonReactStatics from 'hoist-non-react-statics';
 
-import type { ReactRouterV6Config, ReactRouterV6DataRouterConfig } from '../../types';
+import type {
+  ReactRouterV6Config,
+  ReactRouterV6DataRouterConfig,
+  ReactRouterV7Config,
+  ReactRouterV7DataRouterConfig,
+} from '../../types';
 import { ReactRouterVersion } from '../types';
 
 import { FaroRoutes } from './FaroRoutes';
@@ -33,6 +38,22 @@ export function createReactRouterV6DataOptions(
 ): ReactRouterV6DataRouterConfig {
   return {
     version: ReactRouterVersion.V6_data_router,
+    dependencies,
+  };
+}
+
+export function createReactRouterV7Options(dependencies: ReactRouterV6Dependencies): ReactRouterV7Config {
+  return {
+    version: ReactRouterVersion.V7,
+    dependencies,
+  };
+}
+
+export function createReactRouterV7DataOptions(
+  dependencies: ReactRouterV6DataRouterDependencies
+): ReactRouterV7DataRouterConfig {
+  return {
+    version: ReactRouterVersion.V7_data_router,
     dependencies,
   };
 }